c851be4 net: define NodeId as an int64_t (Cory Fields)

Tree-SHA512: 2ccc931cfcdc555313b9434d8de2f6cea759b31891212ca62f962208f60157d4fc593010e3ca61265d1a20d6f78c6ca79103600b85df77983d5509d192875b96

+++ src/net.h
@@ -100,7 +100,7 @@ static const unsigned int DEFAULT_MISBEHAVING_BANTIME = 60 * 60 * 24;  // Defaul
 static const auto DEFAULT_OUTBOUND_INTERVAL = std::chrono::milliseconds(500);
 static const auto MIN_OUTBOUND_INTERVAL = std::chrono::milliseconds(100);
 
-typedef int NodeId;
+typedef int64_t NodeId;
 
 struct AddedNodeInfo
 {
